+++ internal/adapter/confluence/confluence.go
@@ -30,6 +30,7 @@ var _ domain.DocStore = (*Confluence)(nil)
 var _ domain.AssetResolver = (*Confluence)(nil)
 var _ domain.Verifier = (*Confluence)(nil)
 var _ domain.PageShortLinkResolver = (*Confluence)(nil)
+var _ domain.CompletePageSearcher = (*Confluence)(nil)
 
 func (cf *Confluence) ResolveShortPageLink(ctx context.Context, path string) (string, error) {
 	return cf.c.ResolveGET(ctx, path)